fix some bug

+++ app/src/main/java/com/sh/entertainment/fastcharge/data/interactor/BoInteractor.kt
@@ -82,7 +82,9 @@ class BoInteractor(private val ctx: Context) : BaseInteractor() {
             val activityManager = ctx.getSystemService(Context.ACTIVITY_SERVICE) as ActivityManager?
             activityManager?.run {
                 for (packageInfo in packages) {
-                    killBackgroundProcesses(packageInfo.packageName)
+                    if(packageInfo.packageName != ctx.packageName){
+                        killBackgroundProcesses(packageInfo.packageName)
+                    }
                 }
             }
         } catch (e: Exception) {
